game 4 50%

+++ Assets/Scripts/ShowScorePanel.cs
@@ -12,24 +12,24 @@ public class ShowScorePanel : MonoBehaviour
     {
 
 
-
+ 
 
         playerText1.text=PlayerScore1.ToString();
         playerText2.text=PlayerScore2.ToString();
-        PlayerPrefs.SetFloat("player1Score", PlayerPrefs.GetFloat("player1Score", 0f)+PlayerScore1);
-        PlayerPrefs.SetFloat("player2Score", PlayerPrefs.GetFloat("player2Score", 0f)+PlayerScore2);
+        PlayerPrefs.SetInt("player1Score", PlayerPrefs.GetInt("player1Score", 0)+PlayerScore1);
+        PlayerPrefs.SetInt("player2Score", PlayerPrefs.GetInt("player2Score", 0)+PlayerScore2);
         PlayerPrefs.Save();
 
-        playerTotalText1.text=PlayerPrefs.GetFloat("player1Score", 0f).ToString();
-        playerTotalText2.text=PlayerPrefs.GetFloat("player2Score", 0f).ToString();
+        playerTotalText1.text=PlayerPrefs.GetInt("player1Score", 0).ToString();
+        playerTotalText2.text=PlayerPrefs.GetInt("player2Score", 0).ToString();
             // PlayerPrefs.SetString("player1",playerName1);
             // PlayerPrefs.SetString("player2",playerName2);
-            // PlayerPrefs.SetFloat("player1Score", 0f);
-            // PlayerPrefs.SetFloat("player2Score", 0f);
+            // PlayerPrefs.SetInt("player1Score", 0);
+            // PlayerPrefs.SetInt("player2Score", 0);
             // PlayerPrefs.Save();
         CarGameDatabase database = new CarGameDatabase();
-        database.UpdateScoreInfo(PlayerPrefs.GetString("player1"),(int)PlayerPrefs.GetFloat("player1Score", 0f));
-        database.UpdateScoreInfo(PlayerPrefs.GetString("player2"),(int)PlayerPrefs.GetFloat("player2Score", 0f));
+        database.UpdateScoreInfo(PlayerPrefs.GetString("player1"),(int)PlayerPrefs.GetInt("player1Score", 0));
+        database.UpdateScoreInfo(PlayerPrefs.GetString("player2"),(int)PlayerPrefs.GetInt("player2Score", 0));
         playerRank1.text=database.GetRankByName(PlayerPrefs.GetString("player1")).ToString(); 
         playerRank2.text=database.GetRankByName(PlayerPrefs.GetString("player2")).ToString();
         thisPanel.SetActive(true);
